Fix dashboard product count and update product management
