+++ src/app/admin/dashboard/content/page.jsx
@@ -23,6 +23,11 @@ export default function ContentManagement() {
         title: ''
       }
     },
+    menu: {
+      title: '',
+      subtitle: '',
+      products: []
+    },
     about: {
       heading: '',
       story: '',
@@ -65,6 +70,11 @@ export default function ContentManagement() {
               title: ''
             }
           },
+          menu: {
+            title: data.content.menu?.title || 'Tacta Slime Menu',
+            subtitle: data.content.menu?.subtitle || 'Discover our handcrafted slimes available at today\'s market!',
+            products: data.content.menu?.products || []
+          },
           about: {
             heading: data.content.about?.heading || '',
             story: data.content.about?.story || '',
@@ -103,7 +113,11 @@ export default function ContentManagement() {
       // Add all form fields to FormData
       Object.entries(content).forEach(([section, fields]) => {
         Object.entries(fields).forEach(([key, value]) => {
-          formData.append(key, value || '');
+          if (key === 'products') {
+            formData.append(key, JSON.stringify(value));
+          } else {
+            formData.append(key, value || '');
+          }
         });
       });
 
@@ -210,7 +224,7 @@ export default function ContentManagement() {
           {/* Tabs */}
           <div className="border-b border-gray-200">
             <nav className="flex space-x-8 px-6" aria-label="Tabs">
-              {['home', 'about'].map((tab) => (
+              {['home', 'menu', 'about'].map((tab) => (
                 <button
                   key={tab}
                   onClick={() => setActiveTab(tab)}
@@ -248,175 +262,96 @@ export default function ContentManagement() {
                   
                   <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
                     <div>
-                      <label htmlFor="heroTitle" className="block text-sm font-medium text-gray-700 mb-1">
-                        Hero Title
-                      </label>
+                      <label className="block text-sm font-medium text-gray-700">Hero Title</label>
                       <input
                         type="text"
-                        id="heroTitle"
                         value={content.home.heroTitle}
                         onChange={(e) => handleContentChange('home', 'heroTitle', e.target.value)}
-                        className="w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-2 focus:ring-tacta-pink focus:border-tacta-pink"
+                        className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-tacta-pink focus:ring-tacta-pink sm:text-sm"
                       />
                     </div>
                     
                     <div>
-                      <label htmlFor="heroSubtitle" className="block text-sm font-medium text-gray-700 mb-1">
-                        Hero Subtitle
-                      </label>
+                      <label className="block text-sm font-medium text-gray-700">Hero Subtitle</label>
                       <input
                         type="text"
-                        id="heroSubtitle"
                         value={content.home.heroSubtitle}
                         onChange={(e) => handleContentChange('home', 'heroSubtitle', e.target.value)}
-                        className="w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-2 focus:ring-tacta-pink focus:border-tacta-pink"
+                        className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-tacta-pink focus:ring-tacta-pink sm:text-sm"
                       />
                     </div>
                   </div>
-                  
+
                   <div>
-                    <label htmlFor="heroButtonText" className="block text-sm font-medium text-gray-700 mb-1">
-                      Button Text
-                    </label>
+                    <label className="block text-sm font-medium text-gray-700">Hero Button Text</label>
                     <input
                       type="text"
-                      id="heroButtonText"
                       value={content.home.heroButtonText}
                       onChange={(e) => handleContentChange('home', 'heroButtonText', e.target.value)}
-                      className="w-full md:w-1/3 px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-2 focus:ring-tacta-pink focus:border-tacta-pink"
+                      className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-tacta-pink focus:ring-tacta-pink sm:text-sm"
                     />
                   </div>
 
                   <div>
-                    <label className="block text-sm font-medium text-gray-700 mb-1">
-                      Hero Image
-                    </label>
-                    <div className="flex items-center space-x-4">
+                    <label className="block text-sm font-medium text-gray-700">Hero Image</label>
+                    <div className="mt-1 flex items-center space-x-4">
+                      <input
+                        type="file"
+                        accept="image/*"
+                        onChange={handleImageChange}
+                        className="block w-full text-sm text-gray-500 file:mr-4 file:py-2 file:px-4 file:rounded-md file:border-0 file:text-sm file:font-semibold file:bg-tacta-pink-light file:text-tacta-pink hover:file:bg-tacta-pink-dark"
+                      />
                       {content.home.heroImagePath && (
-                        <div className="relative w-32 h-32 rounded-lg overflow-hidden border border-gray-300">
+                        <div className="relative h-20 w-20">
                           <Image
                             src={content.home.heroImagePath}
-                            alt="Hero Image"
+                            alt="Hero Preview"
                             fill
-                            className="object-cover"
+                            className="object-cover rounded-md"
                           />
                         </div>
                       )}
-                      <div className="flex-1">
-                        <input
-                          type="file"
-                          accept="image/*"
-                          onChange={handleImageChange}
-                          className="block w-full text-sm text-gray-500 file:mr-4 file:py-2 file:px-4 file:rounded-md file:border-0 file:text-sm file:font-medium file:bg-tacta-pink-light file:text-tacta-pink hover:file:bg-tacta-pink-light/90"
-                        />
-                        <p className="mt-1 text-sm text-gray-500">
-                          Recommended size: 1200x800px. Max file size: 5MB
-                        </p>
-                      </div>
                     </div>
                   </div>
-                  
-                  <hr className="my-6 border-gray-200" />
-                  
-                  <h3 className="text-lg font-medium text-gray-900">Hero Video</h3>
-                  
-                  <div className="space-y-4">
-                    <div>
-                      <label htmlFor="homeVideoType" className="block text-sm font-medium text-gray-700">
-                        Video Type
-                      </label>
-                      <select
-                        id="homeVideoType"
-                        value={content.home.video.type}
-                        onChange={(e) => handleContentChange('home', 'video', {
-                          ...content.home.video,
-                          type: e.target.value
-                        })}
-                        className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-tacta-pink focus:ring-tacta-pink"
-                      >
-                        <option value="youtube">YouTube</option>
-                        <option value="vimeo">Vimeo</option>
-                        <option value="other">Direct Video URL</option>
-                      </select>
-                    </div>
-                    
-                    <div>
-                      <label htmlFor="homeVideoUrl" className="block text-sm font-medium text-gray-700">
-                        Video URL
-                      </label>
-                      <input
-                        type="text"
-                        id="homeVideoUrl"
-                        value={content.home.video.url}
-                        onChange={(e) => handleContentChange('home', 'video', {
-                          ...content.home.video,
-                          url: e.target.value
-                        })}
-                        placeholder={
-                          content.home.video.type === 'youtube' 
-                            ? 'https://www.youtube.com/watch?v=...' 
-                            : content.home.video.type === 'vimeo'
-                            ? 'https://vimeo.com/...'
-                            : 'https://example.com/video.mp4'
-                        }
-                        className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-tacta-pink focus:ring-tacta-pink"
-                      />
-                      <p className="mt-1 text-sm text-gray-500">
-                        {content.home.video.type === 'youtube' 
-                          ? 'Enter the full YouTube video URL or video ID'
-                          : content.home.video.type === 'vimeo'
-                          ? 'Enter the full Vimeo video URL or video ID'
-                          : 'Enter the direct URL to your video file'}
-                      </p>
-                    </div>
-                    
-                    <div>
-                      <label htmlFor="homeVideoTitle" className="block text-sm font-medium text-gray-700">
-                        Video Title
-                      </label>
-                      <input
-                        type="text"
-                        id="homeVideoTitle"
-                        value={content.home.video.title}
-                        onChange={(e) => handleContentChange('home', 'video', {
-                          ...content.home.video,
-                          title: e.target.value
-                        })}
-                        placeholder="Hero Video"
-                        className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-tacta-pink focus:ring-tacta-pink"
-                      />
-                    </div>
-                  </div>
-                  
-                  <hr className="my-6 border-gray-200" />
-                  
-                  <h3 className="text-lg font-medium text-gray-900">Featured Products Section</h3>
+                </div>
+              )}
+
+              {/* Menu Page Content */}
+              {activeTab === 'menu' && (
+                <div className="space-y-6">
+                  <h3 className="text-lg font-medium text-gray-900">Menu Page Settings</h3>
                   
                   <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
                     <div>
-                      <label htmlFor="featuredTitle" className="block text-sm font-medium text-gray-700 mb-1">
-                        Featured Title
-                      </label>
+                      <label className="block text-sm font-medium text-gray-700">Menu Title</label>
                       <input
                         type="text"
-                        id="featuredTitle"
-                        value={content.home.featuredTitle}
-                        onChange={(e) => handleContentChange('home', 'featuredTitle', e.target.value)}
-                        className="w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-2 focus:ring-tacta-pink focus:border-tacta-pink"
+                        value={content.menu.title}
+                        onChange={(e) => handleContentChange('menu', 'title', e.target.value)}
+                        className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-tacta-pink focus:ring-tacta-pink sm:text-sm"
                       />
                     </div>
                     
                     <div>
-                      <label htmlFor="featuredSubtitle" className="block text-sm font-medium text-gray-700 mb-1">
-                        Featured Subtitle
-                      </label>
+                      <label className="block text-sm font-medium text-gray-700">Menu Subtitle</label>
                       <input
                         type="text"
-                        id="featuredSubtitle"
-                        value={content.home.featuredSubtitle}
-                        onChange={(e) => handleContentChange('home', 'featuredSubtitle', e.target.value)}
-                        className="w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-2 focus:ring-tacta-pink focus:border-tacta-pink"
+                        value={content.menu.subtitle}
+                        onChange={(e) => handleContentChange('menu', 'subtitle', e.target.value)}
+                        className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-tacta-pink focus:ring-tacta-pink sm:text-sm"
                       />
+                    </div>
+                  </div>
+
+                  <div className="border-t border-gray-200 pt-6">
+                    <h4 className="text-lg font-medium text-gray-900 mb-4">Products</h4>
+                    <div className="flex justify-end mb-4">
+                      <Link 
+                        href="/admin/dashboard/products"
+                        className="btn-primary text-sm"
+                      >
+                        Manage Products
+                      </Link>
                     </div>
                   </div>
                 </div>
@@ -425,130 +360,46 @@ export default function ContentManagement() {
               {/* About Page Content */}
               {activeTab === 'about' && (
                 <div className="space-y-6">
-                  <h3 className="text-lg font-medium text-gray-900">About Section</h3>
+                  <h3 className="text-lg font-medium text-gray-900">About Page Content</h3>
                   
                   <div>
-                    <label htmlFor="aboutHeading" className="block text-sm font-medium text-gray-700 mb-1">
-                      Heading
-                    </label>
+                    <label className="block text-sm font-medium text-gray-700">Page Heading</label>
                     <input
                       type="text"
-                      id="aboutHeading"
                       value={content.about.heading}
                       onChange={(e) => handleContentChange('about', 'heading', e.target.value)}
-                      className="w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-2 focus:ring-tacta-pink focus:border-tacta-pink"
-                    />
-                  </div>
-                  
-                  <div>
-                    <label htmlFor="aboutStory" className="block text-sm font-medium text-gray-700 mb-1">
-                      Story
-                    </label>
-                    <textarea
-                      id="aboutStory"
-                      rows={4}
-                      value={content.about.story}
-                      onChange={(e) => handleContentChange('about', 'story', e.target.value)}
-                      className="w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-2 focus:ring-tacta-pink focus:border-tacta-pink"
-                    />
-                  </div>
-                  
-                  <div>
-                    <label htmlFor="missionTitle" className="block text-sm font-medium text-gray-700 mb-1">
-                      Mission Title
-                    </label>
-                    <input
-                      type="text"
-                      id="missionTitle"
-                      value={content.about.missionTitle}
-                      onChange={(e) => handleContentChange('about', 'missionTitle', e.target.value)}
-                      className="w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-2 focus:ring-tacta-pink focus:border-tacta-pink"
-                    />
-                  </div>
-                  
-                  <div>
-                    <label htmlFor="missionText" className="block text-sm font-medium text-gray-700 mb-1">
-                      Mission Text
-                    </label>
-                    <textarea
-                      id="missionText"
-                      rows={4}
-                      value={content.about.missionText}
-                      onChange={(e) => handleContentChange('about', 'missionText', e.target.value)}
-                      className="w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-2 focus:ring-tacta-pink focus:border-tacta-pink"
+                      className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-tacta-pink focus:ring-tacta-pink sm:text-sm"
                     />
                   </div>
 
-                  <hr className="my-6 border-gray-200" />
-                  
-                  <h3 className="text-lg font-medium text-gray-900">About Video</h3>
-                  
-                  <div className="space-y-4">
-                    <div>
-                      <label htmlFor="aboutVideoType" className="block text-sm font-medium text-gray-700">
-                        Video Type
-                      </label>
-                      <select
-                        id="aboutVideoType"
-                        value={content.about.video.type}
-                        onChange={(e) => handleContentChange('about', 'video', {
-                          ...content.about.video,
-                          type: e.target.value
-                        })}
-                        className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-tacta-pink focus:ring-tacta-pink"
-                      >
-                        <option value="youtube">YouTube</option>
-                        <option value="vimeo">Vimeo</option>
-                        <option value="other">Direct Video URL</option>
-                      </select>
-                    </div>
-                    
-                    <div>
-                      <label htmlFor="aboutVideoUrl" className="block text-sm font-medium text-gray-700">
-                        Video URL
-                      </label>
-                      <input
-                        type="text"
-                        id="aboutVideoUrl"
-                        value={content.about.video.url}
-                        onChange={(e) => handleContentChange('about', 'video', {
-                          ...content.about.video,
-                          url: e.target.value
-                        })}
-                        placeholder={
-                          content.about.video.type === 'youtube' 
-                            ? 'https://www.youtube.com/watch?v=...' 
-                            : content.about.video.type === 'vimeo'
-                            ? 'https://vimeo.com/...'
-                            : 'https://example.com/video.mp4'
-                        }
-                        className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-tacta-pink focus:ring-tacta-pink"
-                      />
-                      <p className="mt-1 text-sm text-gray-500">
-                        {content.about.video.type === 'youtube' 
-                          ? 'Enter the full YouTube video URL or video ID'
-                          : content.about.video.type === 'vimeo'
-                          ? 'Enter the full Vimeo video URL or video ID'
-                          : 'Enter the direct URL to your video file'}
-                      </p>
-                    </div>
-                    
-                    <div>
-                      <label htmlFor="aboutVideoTitle" className="block text-sm font-medium text-gray-700">
-                        Video Title
-                      </label>
-                      <input
-                        type="text"
-                        id="aboutVideoTitle"
-                        value={content.about.video.title}
-                        onChange={(e) => handleContentChange('about', 'video', {
-                          ...content.about.video,
-                          title: e.target.value
-                        })}
-                        placeholder="About Video"
-                        className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-tacta-pink focus:ring-tacta-pink"
-                      />
-                    </div>
+                  <div>
+                    <label className="block text-sm font-medium text-gray-700">Our Story</label>
+                    <textarea
+                      rows={4}
+                      value={content.about.story}
+                      onChange={(e) => handleContentChange('about', 'story', e.target.value)}
+                      className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-tacta-pink focus:ring-tacta-pink sm:text-sm"
+                    />
+                  </div>
+
+                  <div>
+                    <label className="block text-sm font-medium text-gray-700">Mission Title</label>
+                    <input
+                      type="text"
+                      value={content.about.missionTitle}
+                      onChange={(e) => handleContentChange('about', 'missionTitle', e.target.value)}
+                      className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-tacta-pink focus:ring-tacta-pink sm:text-sm"
+                    />
+                  </div>
+
+                  <div>
+                    <label className="block text-sm font-medium text-gray-700">Mission Text</label>
+                    <textarea
+                      rows={4}
+                      value={content.about.missionText}
+                      onChange={(e) => handleContentChange('about', 'missionText', e.target.value)}
+                      className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-tacta-pink focus:ring-tacta-pink sm:text-sm"
+                    />
                   </div>
                 </div>
               )}
@@ -557,7 +408,7 @@ export default function ContentManagement() {
                 <button
                   type="submit"
                   disabled={isSubmitting}
-                  className="px-4 py-2 bg-tacta-pink text-white rounded-md hover:bg-tacta-pink/90 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-tacta-pink disabled:opacity-50 disabled:cursor-not-allowed"
+                  className={`btn-primary ${isSubmitting ? 'opacity-50 cursor-not-allowed' : ''}`}
                 >
                   {isSubmitting ? 'Saving...' : 'Save Changes'}
                 </button>
